coaching_main/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Audio Processing
assemblyai>=0.44.0  # Need v3 streaming API support
pyaudio==0.2.14
numpy==1.26.4
websockets>=11.0  # For WebSocket support (if needed)

# AI/ML Models
google-generativeai==0.3.1
torch==2.9.0
torchaudio==2.9.0
transformers==4.35.0

# Database
chromadb==0.4.18

# Frontend
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
websocket-client==1.6.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0